requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
requests>=2.31.0

